added position to all lights
